% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_results_table.R
\name{add_label_analysis_key}
\alias{add_label_analysis_key}
\title{Helper to add the label analysis key}
\usage{
add_label_analysis_key(results)
}
\arguments{
\item{results}{results table with the label_* columns}
}
\value{
results with label analysis key
}
\description{
Helper to add the label analysis key
}
