% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_results_table.R
\name{add_label_columns_to_results_table}
\alias{add_label_columns_to_results_table}
\title{Add labels to the result table}
\usage{
add_label_columns_to_results_table(results_table, dictionary)
}
\arguments{
\item{results_table}{Result object with an analysis key.}

\item{dictionary}{Dictionary created with create_label_dictionary}
}
\value{
results table with label columns
}
\description{
Add labels to the result table
}
\examples{
add_label_columns_to_results_table(
  results_table = presentresults_MSNA2024_results_table,
  dictionary = presentresults_MSNA2024_dictionary
)
}
