% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cluster_and_strata_summary.R
\name{create_group_clusters}
\alias{create_group_clusters}
\title{Create number of cluster and number of hh surveyed per group/strata}
\usage{
create_group_clusters(
  result,
  analysis_key = "analysis_key",
  dataset,
  cluster_name = NULL
)
}
\arguments{
\item{result}{A result table which must contain \code{analysis_key} column.}

\item{analysis_key}{String with the name of the analysis key. Default is "analysis_key"}

\item{dataset}{dataset}

\item{cluster_name}{String with the name of the cluster id in the dataset.}
}
\value{
A dataframe with number of HH and cluster in each group
}
\description{
Create number of cluster and number of hh surveyed per group/strata
}
\examples{
create_group_clusters(
  result = presentresults_resultstable,
  dataset = presentresults_MSNA_template_data,
  cluster_name = "cluster_id"
)

}
