% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_results_table.R
\name{create_label_dictionary}
\alias{create_label_dictionary}
\title{Create a dictionary for the labeling results}
\usage{
create_label_dictionary(
  kobo_survey_sheet,
  kobo_choices_sheet,
  label_column = "label::english",
  analysis_type_dictionary = NULL,
  results_table = NULL
)
}
\arguments{
\item{kobo_survey_sheet}{KOBO survey sheet to be used.}

\item{kobo_choices_sheet}{KOBO choices sheet to be used.}

\item{label_column}{label column from the KOBO tools to be used.}

\item{analysis_type_dictionary}{Analysis type dictionary, a data frame with analysis type and
label_analysis_type to be used. By default parameters is set to NULL. It will use a default
dataframe. See section analysis_type_dictionary for more details.}

\item{results_table}{result object with an analysis key. Default is NULL, it will be used with
review_kobo_labels.}
}
\value{
A list with 3 dataframes: dictionary_survey, dictionary_choices, analysis_type_dictionary
}
\description{
Create a dictionary for the labeling results
}
\section{analysis_type_dictionary}{

The default analysis dictionary is created like that.

\if{html}{\out{<div class="sourceCode">}}\preformatted{data.frame(analysis_type = c("prop_select_one",
                             "prop_select_multiple",
                             "mean",
                             "ratio",
                             "median"),
           label_analysis_type = c("Proportion (single choice)",
                                   "Proportion (multiple choice)",
                                   "Mean",
                                   "Ratio",
                                   "Median")

}\if{html}{\out{</div>}}
}

\examples{

create_label_dictionary(
  kobo_survey_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_survey,
  kobo_choices_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_choices
)


french_dictionary <- data.frame(
  analysis_type = c(
    "prop_select_one",
    "prop_select_multiple",
    "mean",
    "ratio",
    "median"
  ),
  label_analysis_type = c(
    "Proportion (Choix unique)",
    "Proportion (Choix multiple)",
    "Moyenne",
    "Ratio",
    "Médiane"
  )
)

create_label_dictionary(
  kobo_survey_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_survey,
  kobo_choices_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_choices,
  label_column = "label::french",
  analysis_type_dictionary = french_dictionary
)

}
