% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ipc_table.R
\name{create_ordered_ipc_table}
\alias{create_ordered_ipc_table}
\title{Helper to order the ipc results}
\usage{
create_ordered_ipc_table(
  results_table,
  analysis_key = "analysis_key",
  fcs_cat_var = "fsl_fcs_cat",
  fcs_cat_values = c("Poor", "Borderline", "Acceptable"),
  fcs_set = c("fcs_cereal", "fcs_pulses", "fcs_milk", "fcs_meat", "fcs_veg", "fcs_fruit",
    "fcs_oil", "fcs_sugar", "fcs_spices"),
  hhs_cat_var = "fsl_hhs_cat_ipc",
  hhs_cat_values = c("None", "No or Little", "Moderate", "Severe", "Very Severe"),
  hhs_cat_yesno_set = c("fsl_hhs_nofoodhh", "fsl_hhs_sleephungry", "fsl_hhs_alldaynight"),
  hhs_value_yesno_set = c("yes", "no"),
  hhs_cat_freq_set = c("fsl_hhs_nofoodhh_freq", "fsl_hhs_sleephungry_freq",
    "fsl_hhs_alldaynight_freq"),
  hhs_value_freq_set = c("rarely", "sometimes", "often"),
  rcsi_cat_var = "fsl_rcsi_cat",
  rcsi_cat_values = c("No to Low", "Medium", "High"),
  rcsi_set = c("fsl_rcsi_lessquality", "fsl_rcsi_borrow", "fsl_rcsi_mealsize",
    "fsl_rcsi_mealadult", "fsl_rcsi_mealnb"),
  lcsi_cat_var = "fsl_lcsi_cat",
  lcsi_cat_values = c("None", "Stress", "Crisis", "Emergency"),
  lcsi_set = c("fsl_lcsi_stress1", "fsl_lcsi_stress2", "fsl_lcsi_stress3",
    "fsl_lcsi_stress4", "fsl_lcsi_crisis1", "fsl_lcsi_crisis2", "fsl_lcsi_crisis3",
    "fsl_lcsi_emergency1", "fsl_lcsi_emergency2", "fsl_lcsi_emergency3"),
  lcsi_value_set = c("yes", "no_had_no_need", "no_exhausted", "not_applicable"),
  with_hdds = TRUE,
  hdds_cat = "fsl_hdds_cat",
  hdds_cat_values = c("Low", "Medium", "High"),
  hdds_set = c("fsl_hdds_cereals", "fsl_hdds_tubers", "fsl_hdds_veg", "fsl_hdds_fruit",
    "fsl_hdds_meat", "fsl_hdds_eggs", "fsl_hdds_fish", "fsl_hdds_legumes",
    "fsl_hdds_dairy", "fsl_hdds_oil", "fsl_hdds_sugar", "fsl_hdds_condiments"),
  hdds_value_set = c("yes", "no"),
  with_fclc = FALSE,
  fclc_matrix_var = "fclcm_phase",
  fclc_matrix_values = c("Phase 1 FCLC", "Phase 2 FCLC", "Phase 3 FCLC", "Phase 4 FCLC",
    "Phase 5 FCLC"),
  fc_matrix_var = "fsl_fc_phase",
  fc_matrix_values = c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC",
    "Phase 5 FC"),
  other_variables = NULL,
  stat_col = "stat",
  proportion_name = "prop_select_one",
  mean_name = "mean"
)
}
\arguments{
\item{results_table}{results table with analysis key}

\item{analysis_key}{String with the name of the analysis key. Default is "analysis_key"}

\item{fcs_cat_var}{The string with the name of the Food Consumption Score. Default is "fsl_fcs_cat"}

\item{fcs_cat_values}{String with the options of the Food Consumption Score. Default is
c("Poor", "Borderline", "Acceptable")}

\item{fcs_set}{String for the Food Consumption Score questions set. Default is
c("fcs_cereal", "fcs_pulses", "fcs_milk", "fcs_meat", "fcs_veg", "fcs_fruit", "fcs_oil",
"fcs_sugar", "fcs_spices")}

\item{hhs_cat_var}{The string with the name of the Household Hunger Scale. Default is "fsl_hhs_cat_ipc"}

\item{hhs_cat_values}{String with the options of the Household Hunger Scale. Default is
c("None", "No or Little", "Moderate", "Severe", "Very Severe")}

\item{hhs_cat_yesno_set}{String for the Household Hunger Scale yes-no questions set. Default is
c("fsl_hhs_nofoodhh", "fsl_hhs_sleephungry", "fsl_hhs_alldaynight")}

\item{hhs_value_yesno_set}{String for the values of the Household Hunger Scale yes-no questions
set. Default is c("yes", "no")}

\item{hhs_cat_freq_set}{String for the Household Hunger Scale frequency questions set. Default is
c("fsl_hhs_nofoodhh_freq", "fsl_hhs_sleephungry_freq", "fsl_hhs_alldaynight_freq")}

\item{hhs_value_freq_set}{String for the values of the Household Hunger Scale frequency questions
set. Default is c("rarely", "sometimes", "often")}

\item{rcsi_cat_var}{The string with the name of the reduced Coping Strategy Index. Default is "fsl_rcsi_cat"}

\item{rcsi_cat_values}{String with the options of the reduced Coping Strategy Index. Default is
c("No to Low", "Medium", "High")}

\item{rcsi_set}{String for the reduced Coping Strategy Index questions set. Default is
c("fsl_rcsi_lessquality", "fsl_rcsi_borrow", "fsl_rcsi_mealsize", "fsl_rcsi_mealadult", "fsl_rcsi_mealnb")}

\item{lcsi_cat_var}{The string with the name of the Livelihood Coping Strategy Index. Default is
"fsl_lcsi_cat"}

\item{lcsi_cat_values}{String with the options of the Livelihood Coping Strategy Index. Default
is c("None", "Stress", "Crisis", "Emergency")}

\item{lcsi_set}{String for the Livelihood Coping Strategy Index questions set. Default is
c("fsl_lcsi_stress1", "fsl_lcsi_stress2", "fsl_lcsi_stress3", "fsl_lcsi_stress4",
"fsl_lcsi_crisis1", "fsl_lcsi_crisis2", "fsl_lcsi_crisis3", "fsl_lcsi_emergency1",
"fsl_lcsi_emergency2", "fsl_lcsi_emergency3")}

\item{lcsi_value_set}{String for the values of the Livelihood Coping Strategy Index questions
set. Default is c("yes", "no_had_no_need", "no_exhausted", "not_applicable")}

\item{with_hdds}{TRUE or FALSE, whether to include the FCLC and FC values.}

\item{hdds_cat}{String with the name of the Household Dietary Diversity Score. Default is "fsl_hdds_cat"}

\item{hdds_cat_values}{String with the options of the Household Dietary Diversity Score. Default is
c("Low", "Medium", "High")}

\item{hdds_set}{String for the Household Dietary Diversity Score. Default is
c("fsl_hdds_cereals", "fsl_hdds_tubers", "fsl_hdds_veg", "fsl_hdds_fruit", "fsl_hdds_meat",
"fsl_hdds_eggs", "fsl_hdds_fish", "fsl_hdds_legumes", "fsl_hdds_dairy", "fsl_hdds_oil",
"fsl_hdds_sugar", "fsl_hdds_condiments")}

\item{hdds_value_set}{String for the values of the Household Dietary Diversity Score questions
set. Default is c("yes", "no")}

\item{with_fclc}{TRUE or FALSE, whether to include the FCLC and FC values.}

\item{fclc_matrix_var}{String with the name of the food consumption livelihood matrix from FEWSNET.
Default is "fclcm_phase"}

\item{fclc_matrix_values}{String with the options of the food consumption livelihood matrix
Default is c("Phase 1 FCLC", "Phase 2 FCLC", "Phase 3 FCLC", "Phase 4 FCLC", "Phase 5 FCLC")}

\item{fc_matrix_var}{String with the name of the food consumption matrix from FEWSNET.
Default is "fsl_fc_phase"}

\item{fc_matrix_values}{String with the options of the food consumption matrix. Default is
c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC", "Phase 5 FC")
Default is set to FALSE.}

\item{other_variables}{String for the names of other variables to include. Default is NULL}

\item{stat_col}{String for the name of the column with the mean or the proportion. Default is
"stat"}

\item{proportion_name}{String how a proportion is called in the analysis key. Default is
"prop_select_one"}

\item{mean_name}{String how a mean is called in the analysis key. Default is "mean"}
}
\value{
one ordered group x variable table
}
\description{
Helper to order the ipc results
}
\examples{

no_nas_presentresults_resultstable <- presentresults_resultstable \%>\%
  dplyr::filter(!(analysis_type == "prop_select_one" & is.na(analysis_var_value)))
create_ordered_ipc_table(
  results_table = no_nas_presentresults_resultstable,
  fcs_cat_var = "fcs_cat",
  fcs_cat_values = c("low", "medium", "high"),
  fcs_set = c(
    "fs_fcs_cereals_grains_roots_tubers",
    "fs_fcs_beans_nuts",
    "fs_fcs_dairy",
    "fs_fcs_meat_fish_eggs",
    "fs_fcs_vegetables_leaves",
    "fs_fcs_fruit",
    "fs_fcs_oil_fat_butter",
    "fs_fcs_sugar",
    "fs_fcs_condiment"
  ),
  hhs_cat_var = "hhs_cat",
  hhs_cat_values = c("none", "slight", "moderate", "severe", "very_severe"),
  hhs_cat_yesno_set = c("fs_hhs_nofood_yn", "fs_hhs_sleephungry_yn", "fs_hhs_daynoteating_yn"),
  hhs_cat_freq_set = c("fs_hhs_nofood_freq", "fs_hhs_sleephungry_freq", "fs_hhs_daynoteating_freq"),
  hhs_value_freq_set = c("rarely_1_2", "sometimes_3_10", "often_10_times"),
  rcsi_cat_var = "rcsi_cat",
  rcsi_cat_values = c("low", "medium", "high"),
  rcsi_set = c("rCSILessQlty", "rCSIBorrow", "rCSIMealSize", "rCSIMealAdult", "rCSIMealNb"),
  lcsi_cat_var = "lcs_cat",
  lcsi_cat_values = c("none", "stress", "emergency", "crisis"),
  lcsi_set = c(
    "liv_stress_lcsi_1",
    "liv_stress_lcsi_2",
    "liv_stress_lcsi_3",
    "liv_stress_lcsi_4",
    "liv_crisis_lcsi_1",
    "liv_crisis_lcsi_2",
    "liv_crisis_lcsi_3",
    "liv_emerg_lcsi_1",
    "liv_emerg_lcsi_2",
    "liv_emerg_lcsi_3"
  ),
  with_hdds = FALSE,
  with_fclc = TRUE,
  fclc_matrix_var = "fcls_cat",
  fclc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  fc_matrix_var = "fcm_cat",
  fc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  other_variables = c("income_v2_sum", "expenditure_food")
)
}
\keyword{internal}
