% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_for_map.R
\name{create_table_for_map}
\alias{create_table_for_map}
\title{Create a table for MSNA Indicator Maps 1.2 tool}
\usage{
create_table_for_map(
  results_table,
  group_var_value_column = "group_var_value",
  analysis_var_column = "analysis_var",
  stat_column = "stat",
  number_classes = 5
)
}
\arguments{
\item{results_table}{Results table from analysistools filtered for only value per admin.}

\item{group_var_value_column}{Name of the column with the group/dependent variable values.
Default is "group_var_value".}

\item{analysis_var_column}{Name of the column with the analysis/independent variable names.
Default is "analysis_var".}

\item{stat_column}{Name of the column with the stat. Default is "stat".}

\item{number_classes}{Number of classes for the map. It will convert percentages to classes.
Default value is 5.}
}
\value{
A wide table with group variables in rows and the indicators coded in their classes in
the columns.
}
\description{
Create a table for MSNA Indicator Maps 1.2 tool
}
\details{
There can be 5 or 6 classes as follow:

5 classes:
\itemize{
\item 1 : 0
\item 2 : <= 25 \%
\item 3 : <= 50 \%
\item 4 : <= 75 \%
\item 5 : <= 100 \%
\item NA: Anything else
}

6 classes:
\itemize{
\item 1 : 0
\item 2 : <= 20 \%
\item 3 : <= 40 \%
\item 4 : <= 60 \%
\item 5 : <= 80 \%
\item 6 : <= 100 \%
\item NA: Anything else
}
}
\examples{
presentresults::presentresults_MSNA2024_results_table |>
  dplyr::filter(
    analysis_var == "wash_drinking_water_source_cat",
    analysis_var_value == "surface_water",
    group_var == "admin1"
  ) |>
  create_table_for_map()

}
