% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_group_x_variable.R
\name{create_table_group_x_variable}
\alias{create_table_group_x_variable}
\title{Create a wide table with indicators in the columns}
\usage{
create_table_group_x_variable(
  .results,
  analysis_key = "analysis_key",
  value_columns = c("stat", "stat_low", "stat_upp")
)
}
\arguments{
\item{.results}{long table with the stats and analysis key}

\item{analysis_key}{name of the columns of the analysis key, as character vector. Default is
"analysis_key"}

\item{value_columns}{names of the columns with the stats of interest, as character vector.
Default is c("stat", "stat_low", "stat_upp")}
}
\value{
a wide table with grouping variables as rows and analysed variables as columns
}
\description{
Create a wide table with indicators in the columns
}
\examples{
create_table_group_x_variable(presentresults_resultstable, value_columns = "stat")

}
