% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_group_x_variable.R
\name{create_xlsx_group_x_variable}
\alias{create_xlsx_group_x_variable}
\title{Write a table group by variable into Excel}
\usage{
create_xlsx_group_x_variable(
  .table_group_x_variable,
  file_path,
  sheet_name = "table_group_x_variable",
  write_file = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{.table_group_x_variable}{a table create by create_table_group_x_variable}

\item{file_path}{File path, it should contains the file name}

\item{sheet_name}{string with the name of the sheet, default is "table_group_x_variable"}

\item{write_file}{Default is TRUE, it will write the file. If set to FALSE, it will return
a workbook object from openxlsx}

\item{overwrite}{Default is FALSE, it will overwrite the file if set to TRUE.}
}
\value{
An excel file formatted.
}
\description{
Write a table group by variable into Excel
}
\examples{
\dontrun{
presentresults_resultstable \%>\%
  create_table_group_x_variable() \%>\%
  create_xlsx_group_x_variable("mytable.xlsx")
}
}
