% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_variable_x_group.R
\name{create_xlsx_variable_x_group}
\alias{create_xlsx_variable_x_group}
\title{Write a table variable by group into Excel}
\usage{
create_xlsx_variable_x_group(
  table_group_x_variable,
  file_path = NULL,
  table_name = "variable_x_group_table",
  value_columns = c("stat", "stat_low", "stat_upp"),
  total_columns = NULL,
  readme_sheet_name = "readme",
  table_sheet_name = "variable_x_group_table",
  overwrite = FALSE
)
}
\arguments{
\item{table_group_x_variable}{a table create by create_table_variable_x_group}

\item{file_path}{File names and path. Default is NULL which will return a workbook instead of an
excel file.}

\item{table_name}{string with the name of table to write. It will only be used if it is part of a list.
Default is "variable_x_group_table".}

\item{value_columns}{string containing the names of the columns with the stats to export}

\item{total_columns}{string containing the names of the columns with the totals
(n, n_total, n_weighted, n_total_weighted, etc.) to export}

\item{readme_sheet_name}{string with the name of the sheet to write the read me page,
default is "readme"}

\item{table_sheet_name}{string with the name of the sheet to write the table,
default is "variable_x_group_table"}

\item{overwrite}{Default is FALSE, it will overwrite the file if set to TRUE.}
}
\value{
An excel file formatted.
}
\description{
Write a table variable by group into Excel
}
\examples{
\dontrun{
presentresults_resultstable \%>\%
  create_table_variable_x_group() \%>\%
  create_xlsx_variable_x_group(file_path = "mytable.xlsx")
}
}
