% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{impact_colors}
\alias{impact_colors}
\title{IMPACT colors hex code}
\usage{
impact_colors
}
\description{
IMPACT colors hex code
}
\details{
IMPACT colors hex code
}
\keyword{datasets}
