% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{impact_palettes}
\alias{impact_palettes}
\title{IMPACT palettes}
\usage{
impact_palettes
}
\description{
IMPACT palettes
}
\details{
IMPACT palettes hex code
}
\keyword{datasets}
