% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentresults-package.R
\docType{package}
\name{presentresults-package}
\alias{presentresults}
\alias{presentresults-package}
\title{presentresults: Helpers to present results into table or graphs}
\description{
The package provides wrappers to turn results from a long format into wide or graphics. It is based on tidyverse mainly. For results in a long format, it will work with an analysis key which have the following format: analysis_type @/@ dependent_variable ~/~ dependent_variable_value @/@ independent_variable ~/~ independent_variable_value
}
\seealso{
Useful links:
\itemize{
  \item \url{https://impact-initiatives.github.io/presentresults/}
}

}
\author{
\strong{Maintainer}: Yann Say \email{yann.say@impact-initiatives.org} (\href{https://orcid.org/0000-0002-7390-4209}{ORCID})

Authors:
\itemize{
  \item Mehedi Khan \email{mh.khan@reach-initiative.org} (\href{https://orcid.org/0000-0002-4276-4485}{ORCID})
}

}
\keyword{internal}
