% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA2024_dictionary}
\alias{presentresults_MSNA2024_dictionary}
\title{MSNA2024 dictionary}
\usage{
presentresults_MSNA2024_dictionary
}
\description{
Dictionary to be used to label results
}
\details{
An example of dictionary to be used in add_label_columns_to_results_table()
create_label_dictionary(kobo_survey_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_survey,
kobo_choices_sheet = presentresults_MSNA2024_kobotool_fixed$kobo_choices,
results_table = presentresults_MSNA2024_results_table)
}
\keyword{datasets}
