% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA2024_kobotool_fixed}
\alias{presentresults_MSNA2024_kobotool_fixed}
\title{results table example}
\usage{
presentresults_MSNA2024_kobotool_fixed
}
\description{
KOBO Template 2024 V10 without duplicated labels and names.
}
\details{
KOBO template 2024 with composite indicators information without duplicated labels and names.
}
\keyword{datasets}
