% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA2024_kobotool_template}
\alias{presentresults_MSNA2024_kobotool_template}
\title{KOBO template for MSNA 2024 with composite indicators information}
\usage{
presentresults_MSNA2024_kobotool_template
}
\description{
KOBO Template 2024 V10
}
\details{
KOBO template 2024 with composite indicators information
}
\keyword{datasets}
