% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA2024_labelled_results_table}
\alias{presentresults_MSNA2024_labelled_results_table}
\title{MSNA 2024 results table with labels.}
\usage{
presentresults_MSNA2024_labelled_results_table
}
\description{
MSNA 2024 labelled results table
}
\keyword{datasets}
