% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA2024_results_table}
\alias{presentresults_MSNA2024_results_table}
\title{results table example for MSNA 2024 tools}
\usage{
presentresults_MSNA2024_results_table
}
\description{
Results table from MSNA2024 template
}
\keyword{datasets}
