% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_MSNA_template_data}
\alias{presentresults_MSNA_template_data}
\title{MSNA dataset generated by xslxform}
\usage{
presentresults_MSNA_template_data
}
\description{
Datasets for example
}
\examples{
presentresults_MSNA_template_data
presentresults_resultstable
}
\keyword{datasets}
