% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{presentresults_resultstable}
\alias{presentresults_resultstable}
\title{results table example}
\usage{
presentresults_resultstable
}
\description{
results table example
}
\keyword{datasets}
