% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_results_table.R
\name{review_kobo_labels}
\alias{review_kobo_labels}
\title{Review if there are duplication of kobo name and labels in the kobo tools}
\usage{
review_kobo_labels(
  kobo_survey_sheet,
  kobo_choices_sheet,
  label_column = "label::english",
  exclude_type = c("begin_group", "end_group", "beging_repeat", "end_repeat", "note"),
  results_table = NULL
)
}
\arguments{
\item{kobo_survey_sheet}{kobo survey sheet. It must contain type and name.}

\item{kobo_choices_sheet}{kobo choices sheet. It must contain list_name and name.}

\item{label_column}{Column name with the label to be used, default is "label::english"}

\item{exclude_type}{Types to exclude in the review, default is c("begin_group", "end_group", "beging_repeat", "end_repeat", "note")}

\item{results_table}{Results table with group_var and analysis_var columns (names of the different
variables). The table will be used to review only the names and label that will appear.}
}
\value{
A data frame with the duplicated cases and the reasons. Empty if there is no duplication.
}
\description{
Review if there are duplication of kobo name and labels in the kobo tools
}
\examples{
review_kobo_labels(
  kobo_survey_sheet = presentresults_MSNA2024_kobotool_template$kobo_survey,
  kobo_choices_sheet = presentresults_MSNA2024_kobotool_template$kobo_choices,
)
review_kobo_labels(
  kobo_survey_sheet = presentresults_MSNA2024_kobotool_template$kobo_survey,
  kobo_choices_sheet = presentresults_MSNA2024_kobotool_template$kobo_choices,
  results_table = presentresults_MSNA2024_results_table
)
}
