% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_impact}
\alias{theme_impact}
\title{Theme for IMPACT Initiative}
\usage{
theme_impact(initiative = "reach")
}
\arguments{
\item{initiative}{Name of the initiative, should be "reach", "impact" or "agora".}
}
\value{
ggplot2 plot with theme_minimal, bold title and color of the initiative.
}
\description{
It will set:
\itemize{
\item theme_minimal,
\item color of the text in REACH gray (see \code{impact_colors$reach_gray}),
\item Title in bold and with the color of the initiative (reach: \code{impact_colors$red},
impact: \code{impact_colors$blue}, and agora: \code{impact_colors$bordeaux})
}
}
\examples{

data_to_plot <- presentresults::presentresults_MSNA2024_labelled_results_table |>
  dplyr::filter(
    analysis_var == "wash_drinking_water_source_cat",
    group_var == "hoh_gender"
  )

data_to_plot \%>\%
  ggplot2::ggplot() +
    ggplot2::geom_col(ggplot2::aes(x =label_analysis_var_value,
                                   y = stat,
                                   fill = label_group_var_value),
                      position = "dodge") +
    ggplot2::labs(title = stringr::str_wrap(unique(data_to_plot$indicator),50),
                  x = stringr::str_wrap(unique(data_to_plot$label_analysis_var),50),
                  fill = stringr::str_wrap(unique(data_to_plot$label_group_var),20)) +
    theme_impact("reach")
}
