% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_results_table.R
\name{unite_labels}
\alias{unite_labels}
\title{Unite labels columns}
\usage{
unite_labels(key_table)
}
\arguments{
\item{key_table}{a key table built with create_analysis_key_table}
}
\value{
a table with label_analysis_var, label_analysis_var_value, label_group_var, and
label_group_var_value united and with a \%/\% as separator
}
\description{
Unite labels columns
}
