% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_analysis_key}
\alias{verify_analysis_key}
\title{Helper to check the format of the analysis key}
\usage{
verify_analysis_key(analysis_key)
}
\arguments{
\item{analysis_key}{vector of analysis key (not the name of the column)}
}
\value{
It will stop the function if the analysis is not in the correct format
}
\description{
Helper to check the format of the analysis key
}
\examples{
verify_analysis_key(presentresults_resultstable$analysis_key)

}
