% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ipc_table.R
\name{create_ordered_ipc_table}
\alias{create_ordered_ipc_table}
\title{Helper to order the ipc results}
\usage{
create_ordered_ipc_table(
  results_table,
  analysis_key = "analysis_key",
  fclc_matrix_var = "fclc_phase",
  fclc_matrix_values = c("Phase 1 - FCLC", "Phase 2 - FCLC", "Phase 3 - FCLC",
    "Phase 4 - FCLC", "Phase 5 - FCLC"),
  fc_matrix_var = "fc_phase",
  fc_matrix_values = c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC",
    "Phase 5 FC"),
  with_fclc = FALSE,
  fcs_cat_var = "fcs_cat",
  fcs_cat_values = c("Poor", "Borderline", "Acceptable"),
  rcsi_cat_var = "rcsi_cat",
  rcsi_cat_values = c("No to Low", "Medium", "High"),
  lcsi_cat_var = "lcsi_cat",
  lcsi_cat_values = c("None", "Stress", "Crisis", "Emergency"),
  hhs_cat_var = "hhs_cat_ipc",
  hhs_cat_values = c("None", "Little", "Moderate", "Severe", "Very Severe"),
  fcs_set = c("fs_fcs_cereals_grains_roots_tubers", "fs_fcs_beans_nuts", "fs_fcs_dairy",
    "fs_fcs_meat_fish_eggs", "fs_fcs_vegetables_leaves", "fs_fcs_fruit",
    "fs_fcs_oil_fat_butter", "fs_fcs_sugar", "fs_fcs_condiment"),
  rcsi_set = c("rCSILessQlty", "rCSIBorrow", "rCSIMealSize", "rCSIMealAdult",
    "rCSIMealNb"),
  lcsi_set,
  lcsi_value_set = c("yes", "no_had_no_need", "no_exhausted", "not_applicable"),
  with_hdds = TRUE,
  hdds_cat = "fsl_hdds_cat",
  hdds_cat_values = c("Low", "Medium", "High"),
  hdds_set = c("fsl_hdds_cereals", "fsl_hdds_tubers", "fsl_hdds_veg", "fsl_hdds_fruit",
    "fsl_hdds_meat", "fsl_hdds_eggs", "fsl_hdds_fish", "fsl_hdds_legumes",
    "fsl_hdds_dairy", "fsl_hdds_oil", "fsl_hdds_sugar", "fsl_hdds_condiments"),
  hdds_value_set = c("yes", "no"),
  other_variables = NULL,
  stat_col = "stat",
  proportion_name = "prop_select_one",
  mean_name = "mean"
)
}
\arguments{
\item{results_table}{results table with analysis key}

\item{analysis_key}{String with the name of the analysis key. Default is "analysis_key"}

\item{fclc_matrix_var}{String with the name of the food consumption livelihood matrix from FEWSNET.
Default is "fclc_phase"}

\item{fclc_matrix_values}{String with the options of the food consumption livelihood matrix
Default is c("Phase 1 - FCLC", "Phase 2 - FCLC", "Phase 3 - FCLC", "Phase 4 - FCLC", "Phase 5 - FCLC")}

\item{fc_matrix_var}{String with the name of the food consumption matrix from FEWSNET.
Default is "fc_phase"}

\item{fc_matrix_values}{String with the options of the food consumption matrix. Default is
c("Phase 1 FC", "Phase 2 FC", "Phase 3 FC", "Phase 4 FC", "Phase 5 FC")}

\item{with_fclc}{TRUE or FALSE, whether to include the FCLC and FC values.
Default is set to FALSE.}

\item{fcs_cat_var}{The string with the name of the Food Consumption Score. Default is "fcs_cat"}

\item{fcs_cat_values}{String with the options of the Food Consumption Score. Default is
c("Poor", "Borderline", "Acceptable")}

\item{rcsi_cat_var}{The string with the name of the reduced Coping Strategy Index. Default is "rcsi_cat"}

\item{rcsi_cat_values}{String with the options of the reduced Coping Strategy Index. Default is
c("No to Low", "Medium", "High")}

\item{lcsi_cat_var}{The string with the name of the Livelihood Coping Strategy Index. Default is
"lcsi_cat"}

\item{lcsi_cat_values}{String with the options of the Livelihood Coping Strategy Index. Default
is c("None", "Stress", "Crisis", "Emergency")}

\item{hhs_cat_var}{The string with the name of the Household Hunger Scale. Default is "hhs_cat_ipc"}

\item{hhs_cat_values}{String with the options of the Household Hunger Scale. Default is
c("None", "Little", "Moderate", "Severe", "Very Severe")}

\item{fcs_set}{String for the Food Consumption Score questions set. Default is
c("fs_fcs_cereals_grains_roots_tubers", "fs_fcs_beans_nuts", "fs_fcs_dairy",
"fs_fcs_meat_fish_eggs", "fs_fcs_vegetables_leaves", "fs_fcs_fruit", "fs_fcs_oil_fat_butter",
"fs_fcs_sugar", "fs_fcs_condiment")}

\item{rcsi_set}{String for the reduced Coping Strategy Index questions set. Default is
c("No to Low", "Medium", "High")}

\item{lcsi_set}{String for the Livelihood Coping Strategy Index questions set. There is
no default}

\item{lcsi_value_set}{String for the values of the Livelihood Coping Strategy Index questions
set. Default is c("yes", "no_had_no_need", "no_exhausted", "not_applicable")}

\item{with_hdds}{TRUE or FALSE, whether to include the FCLC and FC values.}

\item{hdds_cat}{String with the name of the Household Dietary Diversity Score. Default is "fsl_hdds_cat"}

\item{hdds_cat_values}{String with the options of the Household Dietary Diversity Score. Default is
c("Low", "Medium", "High")}

\item{hdds_set}{String for the Household Dietary Diversity Score. Default is
c("fsl_hdds_cereals", "fsl_hdds_tubers", "fsl_hdds_veg", "fsl_hdds_fruit", "fsl_hdds_meat",
"fsl_hdds_eggs", "fsl_hdds_fish", "fsl_hdds_legumes", "fsl_hdds_dairy", "fsl_hdds_oil",
"fsl_hdds_sugar", "fsl_hdds_condiments")}

\item{hdds_value_set}{String for the values of the Household Dietary Diversity Score questions
set. Default is c("yes", "no")}

\item{other_variables}{String for the names of other variables to include. Default is NULL}

\item{stat_col}{String for the name of the column with the mean or the proportion. Default is
"stat"}

\item{proportion_name}{String how a proportion is called in the analysis key. Default is
"prop_select_one"}

\item{mean_name}{String how a mean is called in the analysis key. Default is "mean"}
}
\value{
one ordered group x variable table
}
\description{
Helper to order the ipc results
}
\examples{
create_ordered_ipc_table(
  results_table = presentresults_resultstable,
  fclc_matrix_var = "fcls_cat",
  fclc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  fc_matrix_var = "fcm_cat",
  fc_matrix_values = c("phase_1", "phase_2", "phase_3", "phase_4", "phase_5"),
  with_fclc = TRUE,
  fcs_cat_values = c("low", "medium", "high"),
  rcsi_cat_values = c("low", "medium", "high"),
  lcsi_cat_var = "lcs_cat",
  lcsi_cat_values = c("none", "stress", "emergency", "crisis"),
  hhs_cat_var = "hhs_cat",
  hhs_cat_values = c("none", "slight", "moderate", "severe", "very_severe"),
  lcsi_set = c(
    "liv_stress_lcsi_1",
    "liv_stress_lcsi_2",
    "liv_stress_lcsi_3",
    "liv_stress_lcsi_4",
    "liv_crisis_lcsi_1",
    "liv_crisis_lcsi_2",
    "liv_crisis_lcsi_3",
    "liv_emerg_lcsi_1",
    "liv_emerg_lcsi_2",
    "liv_emerg_lcsi_3"
  ),
  with_hdds = FALSE,
  other_variables = c("income_v2_sum", "expenditure_food")
)
}
