% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_variable_x_group.R
\name{create_table_variable_x_group}
\alias{create_table_variable_x_group}
\title{Turns a long format table into a wide format}
\usage{
create_table_variable_x_group(
  results,
  analysis_key = "analysis_index",
  value_columns = c("stat", "stat_low", "stat_upp"),
  list_for_excel = FALSE
)
}
\arguments{
\item{results}{dataframe with results in a long format}

\item{analysis_key}{analysis key following this description
"analysis_type @/@ dependent_variable  ~/~ dependent_variable_value @/@
independent_variable ~/~ independent_variable_value "}

\item{value_columns}{string contening the names of the variables with the
stats to export}

\item{list_for_excel}{Default is FALSE, the function will return a dataframe.
If set to TRUE, it will return a list of dataframe with the grouping
variable as name. This format makes it easier to write excel files with
different tab.}
}
\value{
a data frame in a wide format with the analysis type, analysis
variable, analysis variable value and the group variable value as columns.
If list_for_excel is set to TRUE, it will return a list per grouping
variable.
}
\description{
Turns a long format table into a wide format
}
\examples{
presentresults_resultstable \%>\% create_table_variable_x_group("analysis_key", "stat")
}
